Continued User and Session work
Clean up getting Users from the database with optional preloads and
add a small UI touch to display the username if a User is logged in.
Move some data to be added to the Conn rather than the Session so
it's not long lived, and add logout functionality.

Lastly, update the formatter to allow a line length appropriate for
reasonably sized displays :)